Spotify/requirements.txt
Flask
gunicorn
tensorflow
numpy
spotipy
flask-cors
Pillow
